Main and Databases
MainWindow: Changed the observable collection to be a property.  Todo: probably need to change the observable collection into a private field that is accessed by public properties.  I think that it might be a moot point when I hook up the code to a database, so I'm not worried about it right now.
Making the observable collection into a property allowed all of the buttons and methods in the main file to access the collection which allowed the create user button to change the listview.

Databases: messed around with SQL databases a little.  I am struggling to connect an existing database to my c# code.  This is an important part of programming, so I'm glad that I'm working through it.  I haven't decided whether it's easier to create a database in MSSM and then connect it to my code in VS, or if I should make an entirely new database in VS.  I am slowly learning to do both